[IMP] auth_api_key: decouple server_environment, making it optional
