+++ auth_api_key/__manifest__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright 2018 ACSONE SA/NV
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
@@ -11,7 +10,7 @@
     "author": "ACSONE SA/NV,Odoo Community Association (OCA)",
     "website": "https://acsone.eu/",
     "development_status": "Beta",
-    "depends": ["server_environment"],
+    "depends": ["base"],
     "data": ['security/ir.model.access.csv', 'views/auth_api_key.xml'],
     "demo": [],
 }
