+++ auth_api_key/models/auth_api_key.py
@@ -2,15 +2,12 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
 from odoo import api, fields, models, tools, _
-
 from odoo.tools import consteq
-
 from odoo.exceptions import ValidationError, AccessError
 
 
 class AuthApiKey(models.Model):
     _name = "auth.api.key"
-    _inherit = "server.env.mixin"
     _description = "API Key"
 
     name = fields.Char(required=True)
@@ -30,27 +27,6 @@ class AuthApiKey(models.Model):
     _sql_constraints = [
         ("name_uniq", "unique(name)", "Api Key name must be unique.")
     ]
-
-    @api.multi
-    def _server_env_section_name(self):
-        """Name of the section in the configuration files
-
-        We override the default implementation to keep the compatibility
-        with the previous implementation of auth_api_key. The section name
-        into the configuration file must be formatted as
-
-            'api_key_{name}'
-
-        """
-        self.ensure_one()
-        return "api_key_{}".format(self.name)
-
-    @property
-    def _server_env_fields(self):
-        base_fields = super()._server_env_fields
-        api_key_fields = {"key": {}}
-        api_key_fields.update(base_fields)
-        return api_key_fields
 
     @api.model
     def _retrieve_api_key(self, key):
