+++ auth_api_key_environment/__manifest__.py
@@ -1,0 +1,15 @@
+# Copyright 2018 ACSONE SA/NV
+# Copyright 2020 Open Source Integrators
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+{
+    "name": "Auth Api Key Server Environment",
+    "summary": "Use Server Environment for API Keys",
+    "version": "12.0.1.0.0",
+    "license": "LGPL-3",
+    "author": "ACSONE SA/NV,Odoo Community Association (OCA)",
+    "website": "https://github.com/OCA/server-auth",
+    "development_status": "Beta",
+    "depends": ["auth_api_key", "server_environment"],
+    "auto_install": True,  # Glue module
+}
