+++ auth_api_key_environment/models/__init__.py
@@ -1,0 +1,1 @@
+from . import auth_api_key
