+++ auth_api_key_environment/models/auth_api_key.py
@@ -1,0 +1,31 @@
+# Copyright 2018 ACSONE SA/NV
+# Copyright 2020 Open Source Integrators
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+from odoo import api, models
+
+
+class AuthApiKey(models.Model):
+    _name = "auth.api.key"
+    _inherit = ["auth.api.key", "server.env.mixin"]
+
+    @api.multi
+    def _server_env_section_name(self):
+        """Name of the section in the configuration files
+
+        We override the default implementation to keep the compatibility
+        with the previous implementation of auth_api_key. The section name
+        into the configuration file must be formatted as
+
+            'api_key_{name}'
+
+        """
+        self.ensure_one()
+        return "api_key_{}".format(self.name)
+
+    @property
+    def _server_env_fields(self):
+        base_fields = super()._server_env_fields
+        api_key_fields = {"key": {}}
+        api_key_fields.update(base_fields)
+        return api_key_fields
