+++ auth_api_key_environment/readme/CONFIGURE.rst
@@ -1,10 +1,10 @@
 The API key menu is available into Settings > Technical in debug mode.
 By default, when you create an API key, the key is saved into the database.
 
-If the ``server_environment`` module is also installed,
-it is also possible to provide the value of this key via the configuration
-file. This can be very useful to avoid mixing your keys between your various
-environments when restoring databases.
+Using Server Environment, it is also possible to provide the value of this key
+via the configuration file.
+This can be very useful to avoid mixing your keys between your various
+environments when restoring databases.i
 
 All you have to do is to add a new
 section to your configuration file according to the following convention:
