+++ auth_api_key_environment/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module makes the Server Environment features available for API Keys.
+For more information, please refer the corresponding module documentation.
