+++ auth_api_key_environment/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_auth_api_key
